Redenumit tari. py.py in tari py si sters vechiul fisier
