+++ Dockerfile
@@ -1,25 +1,11 @@
-ROM python:3.10-alpine
+FROM python:3.10-slim
 
-# Creăm utilizatorul
-RUN adduser -D tari
+WORKDIR /app
 
-# Ne logăm ca utilizator non-root
-USER tari
+COPY . .
 
-# Setăm directorul de lucru
-WORKDIR /home/tari/
+RUN pip install --no-cache-dir -r quickrequirements.txt
 
-# Copiem toate fișierele necesare
-COPY app/ app/
-COPY main.py main.py
-COPY quickrequirements.txt quickrequirements.txt
+EXPOSE 5000
 
-# Instalăm dependențele
-RUN pip install --upgrade pip
-RUN pip install -r quickrequirements.txt
-
-# Expunem portul pe care va rula aplicația
-EXPOSE 5050
-
-# Pornim aplicația cu uvicorn
-CMD ["uvicorn", "main:api", "--host", "0.0.0.0", "--port", "5050"]
+CMD ["python", "tari.py"]
