+++ quickrequirements.txt
@@ -1,0 +1,5 @@
+flask
+pytest
+pylint
+matplotlib
+gunicorn
