+++ tari.py
@@ -1,31 +1,27 @@
-from flask import Flask, url_for
-from app.lib import biblioteca_header, biblioteca_romania, biblioteca_germania, biblioteca_olanda, biblioteca_franta
+from flask import Flask
 
-api = Flask(__name__)
+app = Flask(__name__)
 
-@api.route("/", methods=['GET'])
-def index() -> str:
-    text = biblioteca_header.header_descriere()
-    text += biblioteca_romania.descriere_romania()
-    text += biblioteca_germania.descriere_germania()
-    text += biblioteca_olanda.descriere_olanda()
-    text += biblioteca_franta.descriere_franta()
-    return text
+def descriere():
+    return "<h1>Norvegia este o țară nordică cunoscută pentru fiorduri și peisaje montane.</h1>"
 
-@api.route("/capitala", methods=['GET'])
-def capitala() -> str:
-    text = biblioteca_header.header_capitala()
-    text += biblioteca_romania.capitala_romania()
-    text += biblioteca_germania.capitala_germania()
-    text += biblioteca_olanda.capitala_olanda()
-    text += biblioteca_franta.capitala_franta()
-    return text
+def capitala():
+    return "<h1>Capitala Norvegiei este Oslo.</h1>"
 
-@api.route("/steag", methods=['GET'])
-def steag() -> str:
-    text = biblioteca_header.header_steag()
-    text += biblioteca_romania.steag_romania()
-    text += biblioteca_germania.steag_germania()
-    text += biblioteca_olanda.steag_olanda()
-    text += biblioteca_franta.steag_franta()
-    return text
+def steag():
+    return "<img src='https://upload.wikimedia.org/wikipedia/commons/d/d9/Flag_of_Norway.svg' width='300'>"
+
+@app.route('/Norvegia')
+def ruta_descriere():
+    return descriere()
+
+@app.route('/Norvegia/capitala')
+def ruta_capitala():
+    return capitala()
+
+@app.route('/Norvegia/steag')
+def ruta_steag():
+    return steag()
+
+if __name__ == '__main__':
+    app.run(debug=True, host="0.0.0.0", port=5000)
